psychologist_bot/pyproject.toml
[project]
name = "psychologist-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "mastodon-py>=2.0.1",
    "python-dotenv>=1.1.0",
]
